+++ pyproject.toml
@@ -15,7 +15,6 @@ authors = [{name = "Francesco Bruzzesi"}]
 dependencies = [
     "numpy",
     "narwhals>=0.7.15",
-    "scikit-learn>=0.19",
     "typing-extensions>=4.4.0; python_version < '3.11'",
 ]
 
@@ -39,6 +38,7 @@ issue-tracker = "https://github.com/fbruzzesi/timebasedcv/issues"
 [project.optional-dependencies]
 polars = ["polars>=0.20.3"]
 pandas = ["pandas>=1.2.0"]
+scikit-learn = ["scikit-learn>=0.19",]
 
 dev = [
     "pre-commit==2.21.0",
@@ -100,6 +100,9 @@ convention = "google"
 
 [tool.ruff.lint.pyupgrade]
 keep-runtime-typing = true
+
+[tool.ruff.format]
+docstring-code-format = true
 
 [tool.interrogate]
 ignore-nested-functions = true
