+++ timebasedcv/__init__.py
@@ -4,7 +4,6 @@ from timebasedcv.splitstate import SplitState
 from timebasedcv.timebasedsplit import (
     ExpandingTimeSplit,
     RollingTimeSplit,
-    TimeBasedCVSplitter,
     TimeBasedSplit,
     _CoreTimeBasedSplit,
 )
@@ -17,6 +16,5 @@ __all__ = (
     "TimeBasedSplit",
     "ExpandingTimeSplit",
     "RollingTimeSplit",
-    "TimeBasedCVSplitter",
     "_CoreTimeBasedSplit",
 )
